+++ kong/db/schema/others/migrations.lua
@@ -5,17 +5,36 @@
 -- at https://konghq.com/enterprisesoftwarelicense/.
 -- [ END OF LICENSE 0867164ffc95e54f04670b5169c09574bdbd9bba ]
 
-local strat_migration = {
-  { up = { type = "string", required = true, len_min = 0 } },
-  { teardown = { type = "function" } },
-}
-
-
 return {
   name = "migration",
   fields = {
     { name      = { type = "string", required = true } },
-    { postgres  = { type = "record", required = true, fields = strat_migration } },
-    { cassandra = { type = "record", required = true, fields = strat_migration } },
+    {
+      postgres  = {
+        type = "record", required = true,
+        fields = {
+          { up = { type = "string", len_min = 0 } },
+          { teardown = { type = "function" } },
+        },
+      },
+    },
+    {
+      cassandra = {
+        type = "record", required = true,
+        fields = {
+          { up = { type = "string", len_min = 0 } },
+          { up_f = { type = "function" } },
+          { teardown = { type = "function" } },
+        },
+      }
+    },
+  },
+  entity_checks = {
+    {
+      at_least_one_of = {
+        "postgres.up", "postgres.teardown",
+        "cassandra.up", "cassandra.up_f", "cassandra.teardown"
+      },
+    },
   },
 }