+++ kong/tools/utils.lua
@@ -95,28 +95,6 @@ _M.pack = function(...) return {n = select("#", ...), ...} end
 -- Explicitly honors the `n` field if given in the table, so it is `nil` safe
 _M.unpack = function(t, i, j) return unpack(t, i or 1, j or t.n or #t) end
 
---- Retrieves the hostname of the local machine
--- @return string  The hostname
-function _M.get_hostname()
-  local result
-  local SIZE = 128
-
-  local buf = ffi_new("unsigned char[?]", SIZE)
-  local res = C.gethostname(buf, SIZE)
-
-  if res == 0 then
-    local hostname = ffi_str(buf, SIZE)
-    result = gsub(hostname, "%z+$", "")
-  else
-    local f = io.popen("/bin/hostname")
-    local hostname = f:read("*a") or ""
-    f:close()
-    result = gsub(hostname, "\n$", "")
-  end
-
-  return result
-end
-
 do
   local _system_infos
 
@@ -125,9 +103,7 @@ do
       return _system_infos
     end
 
-    _system_infos = {
-      hostname = _M.get_hostname()
-    }
+    _system_infos = {}
 
     local ok, _, stdout = pl_utils.executeex("getconf _NPROCESSORS_ONLN")
     if ok then