+++ spec/01-unit/05-utils_spec.lua
@@ -10,19 +10,11 @@ local pl_path = require "pl.path"
 
 describe("Utils", function()
 
-  describe("get_hostname()", function()
-    it("should retrieve the hostname", function()
-      assert.is_string(utils.get_hostname())
-    end)
-  end)
-
   describe("get_system_infos()", function()
     it("retrieves various host infos", function()
       local infos = utils.get_system_infos()
       assert.is_number(infos.cores)
-      assert.is_string(infos.hostname)
       assert.is_string(infos.uname)
-      assert.not_matches("\n$", infos.hostname)
       assert.not_matches("\n$", infos.uname)
     end)
     it("caches the result", function()