+++ spec/03-plugins/01-tcp-log/01-tcp-log_spec.lua
@@ -105,6 +105,29 @@ for _, strategy in helpers.each_strategy() do
         },
       }
 
+      local route5 = bp.routes:insert {
+        hosts = { "early_termination.example.com" },
+      }
+
+      bp.plugins:insert {
+        route = { id = route5.id },
+        name     = "tcp-log",
+        config   = {
+          host   = "127.0.0.1",
+          port   = TCP_PORT,
+        },
+      }
+
+      bp.plugins:insert {
+        name   = "request-termination",
+        route  = { id = route5.id },
+        config = {
+          status_code  = 200,
+          content_type = "text/plain",
+          body         = "hello!",
+        },
+      }
+
       local tcp_srv = bp.services:insert({
         name = "tcp",
         host = helpers.mock_upstream_host,
@@ -446,9 +469,30 @@ for _, strategy in helpers.each_strategy() do
       local log_message = cjson.decode(res)
 
       assert.equal("grpcs", log_message.service.protocol)
-      assert.equal("TLSv1.2", log_message.request.tls.version)
+      assert.equal("TLSv1.3", log_message.request.tls.version)
       assert.is_string(log_message.request.tls.cipher)
       assert.equal("NONE", log_message.request.tls.client_verify)
+    end)
+
+    it("tries field encoded as JSON array instead of object #6390", function()
+      local thread = helpers.tcp_server(TCP_PORT) -- Starting the mock TCP server
+
+      -- Making the request
+      local r = assert(proxy_client:send {
+        method  = "GET",
+        path    = "/request",
+        headers = {
+          host  = "early_termination.example.com",
+        },
+      })
+      assert.response(r).has.status(200)
+
+      -- Getting back the TCP server input
+      local ok, res = thread:join()
+      assert.True(ok)
+      assert.is_string(res)
+
+      assert.matches('"tries":[]', res, nil, true)
     end)
 
     it("#stream reports tcp streams", function()