+++ spec/fixtures/custom_plugins/kong/plugins/error-generator-last/handler.lua
@@ -5,31 +5,21 @@
 -- at https://konghq.com/enterprisesoftwarelicense/.
 -- [ END OF LICENSE 0867164ffc95e54f04670b5169c09574bdbd9bba ]
 
-local BasePlugin = require "kong.plugins.base_plugin"
-
 
 local error = error
 
 
-local ErrorGeneratorLastHandler = BasePlugin:extend()
+local ErrorGeneratorLastHandler = {}
 
 
 ErrorGeneratorLastHandler.PRIORITY = -math.huge
 
 
-function ErrorGeneratorLastHandler:new()
-  ErrorGeneratorLastHandler.super.new(self, "error-generator-last")
-end
-
-
 function ErrorGeneratorLastHandler:init_worker()
-  ErrorGeneratorLastHandler.super.init_worker(self)
 end
 
 
 function ErrorGeneratorLastHandler:certificate(conf)
-  ErrorGeneratorLastHandler.super.certificate(self)
-
   if conf.certificate then
     error("[error-generator-last] certificate")
   end
@@ -37,8 +27,6 @@ end
 
 
 function ErrorGeneratorLastHandler:rewrite(conf)
-  ErrorGeneratorLastHandler.super.rewrite(self)
-
   if conf.rewrite then
     error("[error-generator-last] rewrite")
   end
@@ -46,8 +34,6 @@ end
 
 
 function ErrorGeneratorLastHandler:preread(conf)
-  ErrorGeneratorLastHandler.super.preread(self)
-
   if conf.preread then
     error("[error-generator-last] preread")
   end
@@ -56,8 +42,6 @@ end
 
 
 function ErrorGeneratorLastHandler:access(conf)
-  ErrorGeneratorLastHandler.super.access(self)
-
   if conf.access then
     error("[error-generator-last] access")
   end
@@ -65,8 +49,6 @@ end
 
 
 function ErrorGeneratorLastHandler:header_filter(conf)
-  ErrorGeneratorLastHandler.super.header_filter(self)
-
   if conf.header_filter then
     error("[error-generator-last] header_filter")
   end
@@ -74,8 +56,6 @@ end
 
 
 function ErrorGeneratorLastHandler:body_filter(conf)
-  ErrorGeneratorLastHandler.super.body_filter(self)
-
   if conf.header_filter then
     error("[error-generator-last] body_filter")
   end
@@ -83,8 +63,6 @@ end
 
 
 function ErrorGeneratorLastHandler:log(conf)
-  ErrorGeneratorLastHandler.super.log(self)
-
   if conf.log then
     error("[error-generator] body_filter")
   end