+++ spec/fixtures/custom_plugins/kong/plugins/error-generator/schema.lua
@@ -5,14 +5,26 @@
 -- at https://konghq.com/enterprisesoftwarelicense/.
 -- [ END OF LICENSE 0867164ffc95e54f04670b5169c09574bdbd9bba ]
 
+local typedefs = require "kong.db.schema.typedefs"
+
+
 return {
+  name = "error-generator",
   fields = {
-    certificate   = { type = "boolean", required = false, default = false },
-    rewrite       = { type = "boolean", required = false, default = false },
-    preread       = { type = "boolean", required = false, default = false },
-    access        = { type = "boolean", required = false, default = false },
-    header_filter = { type = "boolean", required = false, default = false },
-    body_filter   = { type = "boolean", required = false, default = false },
-    log           = { type = "boolean", required = false, default = false },
+    { consumer = typedefs.no_consumer },
+    { protocols = typedefs.protocols_http },
+    { config = {
+        type = "record",
+        fields = {
+          { certificate   = { type = "boolean", required = false, default = false }, },
+          { rewrite       = { type = "boolean", required = false, default = false }, },
+          { preread       = { type = "boolean", required = false, default = false }, },
+          { access        = { type = "boolean", required = false, default = false }, },
+          { header_filter = { type = "boolean", required = false, default = false }, },
+          { body_filter   = { type = "boolean", required = false, default = false }, },
+          { log           = { type = "boolean", required = false, default = false }, },
+        },
+      }
+    }
   }
 }